lg69t/requirements.txt
https://github.com/PointOneNav/fusion-engine-client/archive/af10b5ae63cd18a308dcb693f4c020bf66d3242b.zip#egg=fusion-engine-client&subdirectory=python
pyserial